+++ yaqd-core/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Changed
+- Updated to fastavro 0.24.0 named schema behavior
+
 ## [2020.07.3]
 
 ### Added